WIP DiscordJS update + mobparts update

+++ index.js
@@ -3,7 +3,15 @@ var stringSimilarity = require('string-similarity');
 var fs = require('fs')
 var path = require('path')
 
-const Bot = new Discord.Client();
+const Bot = new Discord.Client({
+    intents: [
+        Discord.Intents.FLAGS.DIRECT_MESSAGES,
+        Discord.Intents.FLAGS.DIRECT_MESSAGE_REACTIONS,
+        Discord.Intents.FLAGS.GUILD_MESSAGES,
+        Discord.Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
+        //Discord.Intents.FLAGS.GUILD_MEMBERS,
+    ]
+});
 var TOKEN = process.env.token;
 
 const messageAwaiters = [];
@@ -70,7 +78,107 @@ function relocateMessage(user, channel, trigger_member) {
 }
 
 const Commands = {
-    async mobparts(msg, args) {
+    async mobparts(message, args) {
+
+
+
+		const data = await new Promise((resolve, reject) => {
+            request(`https://www.wynem.com/bot_assets/json/cem_template_models.json`, async (err, res, body) => {
+                if (err) {
+                    reject(err);
+                    return;
+                }
+                let d = JSON.parse(body);
+                resolve(d)
+            });
+        })
+		const supported = data.categories.find(e => e.name === "Supported").entities
+		const legacy = data.categories.find(e => e.name === "Legacy")?.entities
+		const unsupported = data.categories.find(e => e.name === "Unsupported")?.entities ?? []
+		const unreleased = data.categories.find(e => e.name === "Unreleased")?.entities ?? []
+		const supportedEntities = supported.map(e => e.name ?? e)
+		const legacyEntities = legacy.map(e => e.name ?? e)
+		const unsupportedEntities = unsupported.map(e => e.name ?? e)
+		const unreleasedEntities = unreleased.map(e => e.name ?? e)
+		const embed = new Discord.MessageEmbed()
+			.setColor("#3CB3FA")
+			.setFooter("You can use the CEM Template Loader plugin to load a working template with the correct part names and pivot points.")
+		if (!args[0]) {
+			embed.setTitle("OptiFine Custom Entity Models")
+					 .setDescription(`Use \`${Bot.getPrefix(message.guild)}cem [entity]\` to view the parts of an individual entity\n\n**Supported**\n\`${supportedEntities.join("` `")}\``)
+					 .addField("Legacy", `\`${legacyEntities.join("` `")}\``)
+			if (unsupportedEntities.length) embed.addField("Unsupported", `\`${unsupportedEntities.join("` `")}\``)
+			if (unreleasedEntities.length) embed.addField("Unreleased", `\`${unreleasedEntities.join("` `")}\``)
+			const row = new Discord.MessageActionRow()
+				.addComponents(
+					new Discord.MessageButton()
+						.setLabel("View online")
+						.setURL("https://www.wynem.com/?cem")
+						.setStyle("LINK"),
+					new Discord.MessageButton()
+						.setLabel("CEM Template Loader")
+						.setURL("https://www.blockbench.net/plugins/cem_template_loader")
+						.setStyle("LINK")
+				)
+			return message.reply({
+				embeds: [embed],
+				allowedMentions: {repliedUser: false},
+				components: [row]
+			}).catch(() => message.channel.send({
+				embeds: [embed],
+				components: [row]
+			}))
+		}
+		args[0] = args[0].toLowerCase().replace(/\s/g, "_")
+		let entityData
+		if (supportedEntities.includes(args[0])) {
+			entityData = supported.find(e => (e.name ?? e) === args[0])
+		} else if (legacyEntities.includes(args[0])) {
+			entityData = legacy.find(e => (e.name ?? e) === args[0])
+		} else if (unsupportedEntities.includes(args[0])) {
+			entityData = unsupported.find(e => (e.name ?? e) === args[0])
+			embed.setFooter("This entity is currently unsupported by OptiFine")
+		} else if (unreleasedEntities.includes(args[0])) {
+			entityData = unreleased.find(e => (e.name ?? e) === args[0])
+			embed.setFooter("This entity is currently unreleased")
+		} else{
+			embed.setTitle(`The entity \`${args[0].substring(0, 128)}\` was not found`)
+            let guess = stringSimilarity.findBestMatch(args[0], supportedEntities.concat(legacyEntities, unsupportedEntities, unreleasedEntities))
+			embed.setDescription(`Did you mean \`${guess}\`?`)
+			embed.setFooter("")
+			return message.reply({
+				embeds: [embed],
+				allowedMentions: {repliedUser: false}
+			}).catch(() => message.channel.send({
+				embeds: [embed]
+			}))
+		}
+		const bones = JSON.parse(data.models[entityData.model ?? entityData.name ?? entityData].model).models.map(e => e.part)
+		const entityName = entityData.display_name ?? (entityData.name ?? entityData).replace(/_/g, " ").toTitleCase()
+		embed.setTitle(entityName)
+		embed.setDescription(`\`${bones.join("` `")}\``)
+		embed.setThumbnail(`https://www.wynem.com/bot_assets/images/minecraft/renders/${args[0]}.png`)
+		const row = new Discord.MessageActionRow()
+			.addComponents(
+				new Discord.MessageButton()
+					.setLabel("View online")
+					.setURL(`https://www.wynem.com/?cem=${args[0]}`)
+					.setStyle("LINK"),
+				new Discord.MessageButton()
+					.setLabel("Download template")
+					.setURL(`https://www.wynem.com/?cem=${args[0]}&download`)
+					.setStyle("LINK")
+			)
+		return message.reply({
+			embeds: [embed],
+			allowedMentions: {repliedUser: false},
+			components: [row]
+		}).catch(() => message.channel.send({
+			embeds: [embed],
+			components: [row]
+		}))
+
+        /*
         let plugin_note = `***Note: **You can use the **CEM Template Loader** plugin to create a working template with the correct part names.*\n`
 
         let file = await new Promise((resolve, reject) => {
@@ -102,7 +210,7 @@ const Commands = {
         } else {
             msg.channel.send(`${plugin_note}OptiFine support the following entities: \n\`\`\`${Object.keys(File.entities.supported).join(', ')}\`\`\``
                 +`\nThese things can currently not be changed: \n\`\`\`${File.entities.unsupported.join(', ')}\`\`\``);
-        }
+        }*/
     },
     /**
      * Runs the FAQ command
@@ -201,6 +309,8 @@ Bot.on('message', msg => {
         return;
     }
 
+    if (msg.channel.type != 'dm' || msg.channel.recipient.username != 'Jannis') return;
+
     if (msg.content.includes('@everyone') && !msg.member.roles.cache.find(role => role.name == 'Moderator')) {
         msg.delete();
         msg.member.kick('Attempting to spam');
@@ -233,6 +343,9 @@ Bot.on('message', msg => {
         }
     }
 
+    if (msg.content.includes('@Moderator')) {
+        log_channel.send(`${msg.author} pinged moderators in ${msg.channel}.`)
+    }
     
     for (var awaiter of messageAwaiters) {
         if ((msg.channel.type == awaiter.channeltype || !awaiter.channeltype)
@@ -266,7 +379,7 @@ Bot.on('message', msg => {
                     + (description ? (description + '\n') : '') + `by ${msg.author}`,
                 image: msg.attachments.first()
             });
-            archive_channel.send(embed).then(pin => {
+            archive_channel.send({embeds: [embed]}).then(pin => {
                 const reactionEmoji = pin.guild.emojis.cache.find(emoji => emoji.name === 'bblike');
                 pin.react(reactionEmoji);
